CHANGE: Bugfixes in number format // php 7.1

+++ MpdfTag.php
@@ -1886,7 +1886,7 @@ class MpdfTag
 					// Don't allow overlap - if floats present - adjust padding to avoid overlap with Floats
 					list($l_exists, $r_exists, $l_max, $r_max, $l_width, $r_width) = $this->mpdf->GetFloatDivInfo($this->mpdf->blklvl - 1);
 					$maxw = $container_w - $l_width - $r_width;
-					if (($setwidth + $currblk['margin_left'] + $currblk['margin_right'] + $bdl + $pdl + $bdr + $pdr) > $maxw || ($maxw - ($currblk['margin_right'] + $currblk['margin_left'] + $bdl + $pdl + $bdr + $pdr)) < (2 * $this->mpdf->GetCharWidth('W', false))) {
+					if (((float) $setwidth + (float) $currblk['margin_left'] + (float) $currblk['margin_right'] + (float) $bdl + (float) $pdl + (float) $bdr + (float) $pdr) > (float) $maxw || ((float) $maxw - (float) ($currblk['margin_right'] + $currblk['margin_left'] + (float) $bdl + (float) $pdl + (float) $bdr + (float) $pdr)) < (2 * $this->mpdf->GetCharWidth('W', false))) {
 						// Too narrow to fit - try to move down past L or R float
 						if ($l_max < $r_max && ($setwidth + $currblk['margin_left'] + $currblk['margin_right'] + $bdl + $pdl + $bdr + $pdr) <= ($container_w - $r_width) && (($container_w - $r_width) - ($currblk['margin_right'] + $currblk['margin_left'] + $bdl + $pdl + $bdr + $pdr)) > (2 * $this->mpdf->GetCharWidth('W', false))) {
 							$this->mpdf->ClearFloats('LEFT', $this->mpdf->blklvl - 1);
@@ -1999,11 +1999,11 @@ class MpdfTag
 					}
 				}
 
-				$currblk['outer_left_margin'] = $prevblk['outer_left_margin'] + $currblk['margin_left'] + $prevblk['border_left']['w'] + $prevblk['padding_left'];
-				$currblk['outer_right_margin'] = $prevblk['outer_right_margin'] + $currblk['margin_right'] + $prevblk['border_right']['w'] + $prevblk['padding_right'];
+				$currblk['outer_left_margin'] = (float) $prevblk['outer_left_margin'] + (float) $currblk['margin_left'] + (float) $prevblk['border_left']['w'] + (float) $prevblk['padding_left'];
+				$currblk['outer_right_margin'] = (float) $prevblk['outer_right_margin'] + (float) $currblk['margin_right'] + (float) $prevblk['border_right']['w'] + (float) $prevblk['padding_right'];
 
-				$currblk['width'] = $this->mpdf->pgwidth - ($currblk['outer_right_margin'] + $currblk['outer_left_margin']);
-				$currblk['inner_width'] = $currblk['width'] - ($currblk['border_left']['w'] + $currblk['padding_left'] + $currblk['border_right']['w'] + $currblk['padding_right']);
+				$currblk['width'] = (float) $this->mpdf->pgwidth - (float) ($currblk['outer_right_margin'] + $currblk['outer_left_margin']);
+				$currblk['inner_width'] = (float) $currblk['width'] - (float) ($currblk['border_left']['w'] + (float) $currblk['padding_left'] + (float) $currblk['border_right']['w'] + (float) $currblk['padding_right']);
 
 				// Check DIV is not now too narrow to fit text
 				$mw = 2 * $this->mpdf->GetCharWidth('W', false);
